requirements.txt
numpy==1.24.3
torch==1.13.1
Pillow==9.5.0
onnxruntime==1.14.1
